Refactor functional bridging tests
