+++ DentOS_Framework/DentOsTestbed/src/dent_os_testbed/test/test_suite/functional/bridging/test_bridging_mac_table_size.py
@@ -6,13 +6,13 @@ from dent_os_testbed.lib.ip.ip_link import IpLink
 
 from dent_os_testbed.utils.test_utils.tgen_utils import (
     tgen_utils_get_dent_devices_with_tgen,
+    tgen_utils_traffic_generator_connect,
+    tgen_utils_dev_groups_from_config,
     tgen_utils_get_traffic_stats,
     tgen_utils_setup_streams,
     tgen_utils_start_traffic,
     tgen_utils_stop_traffic,
-    tgen_utils_get_loss,
-    tgen_utils_dev_groups_from_config,
-    tgen_utils_traffic_generator_connect,
+    tgen_utils_get_loss
 )
 
 pytestmark = [
@@ -89,12 +89,10 @@ async def test_bridging_mac_table_size(testbed):
         'streamA': {
             'ip_source': dev_groups[tg_ports[0]][0]['name'],
             'ip_destination': dev_groups[tg_ports[1]][0]['name'],
-            'srcMac': {
-                'type': 'increment',
-                'start': '00:00:00:00:00:35',
-                'step': '00:00:00:00:10:00',
-                'count': pps_value
-            },
+            'srcMac': {'type': 'increment',
+                       'start': '00:00:00:00:00:35',
+                       'step': '00:00:00:00:10:00',
+                       'count': pps_value},
             'dstMac': 'aa:bb:cc:dd:ee:11',
             'type': 'raw',
             'protocol': '802.1Q',
@@ -114,8 +112,8 @@ async def test_bridging_mac_table_size(testbed):
         loss = tgen_utils_get_loss(row)
         assert loss == 0, f'Expected loss: 0%, actual: {loss}%'
 
-    rc, out = await dent_dev.run_cmd("bridge fdb show br br0   |  grep 'extern_learn.*offload'  |  wc -l")
-    assert rc == 0, "Failed to grep 'extern_learn.*offload'.\n"
+    rc, out = await dent_dev.run_cmd("bridge fdb show br br0 | grep 'extern_learn.*offload' | wc -l")
+    assert rc == 0, "Failed to grep 'extern_learn.*offload'."
 
     amount = int(out) - ixia_vhost_mac_count
     err_msg = f'Expected count of extern_learn offload entities: 4000, Actual count: {amount}'
